rename php -> setting
