+++ test/src/cgi/01_test.cpp
@@ -1,6 +1,6 @@
 #include "../../../inc/cgi.hpp"
 
-#include "php.hpp"
+#include "setting.hpp"
 
 int
 test_cgi_exec(void)
